requirements.txt

numpy
streamlit
openai
python-dotenv
requests
uuid
faker
fastapi
uvicorn
mlflow
pymssql
pyodbc
faker
pandas
openpyxl
xlrd

# Additional packages for modular Text2SQL API
pydantic==2.5.0
pydantic-settings==2.1.0
chromadb==0.4.18
llama-index==0.9.8
llama-index-vector-stores-chroma==0.1.3
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
ipython==8.17.2
chromadb

# azure-identity 
# azure-keyvault-secrets 
# azure-storage-blob
# azure-search-documents
# pip install azure-ai-projects azure-identity
# ! pip install azure-monitor-opentelemetry opentelemetry-sdk

# Register the virtual environment as a Jupyter kernel:
# python -m ipykernel install --user --name=.venv --display-name="Python (.venv)"
